+++ db-scheduler/src/main/java/com/github/kagkarlsson/scheduler/SchedulerClient.java
@@ -17,6 +17,7 @@ package com.github.kagkarlsson.scheduler;
 
 import com.github.kagkarlsson.scheduler.jdbc.DefaultJdbcCustomization;
 import com.github.kagkarlsson.scheduler.jdbc.JdbcCustomization;
+import com.github.kagkarlsson.scheduler.jdbc.JdbcTaskRepository;
 import com.github.kagkarlsson.scheduler.stats.StatsRegistry;
 import com.github.kagkarlsson.scheduler.task.Execution;
 import com.github.kagkarlsson.scheduler.task.Task;
@@ -170,7 +171,7 @@ public interface SchedulerClient {
         public static Builder create(DataSource dataSource, List<Task<?>> knownTasks) {
             return new Builder(dataSource, knownTasks);
         }
-        
+
         public Builder serializer(Serializer serializer) {
             this.serializer = serializer;
             return this;
