+++ db-scheduler/src/main/java/com/github/kagkarlsson/scheduler/jdbc/JdbcTaskRepository.java
@@ -13,16 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.kagkarlsson.scheduler;
+package com.github.kagkarlsson.scheduler.jdbc;
 
 import com.github.kagkarlsson.jdbc.JdbcRunner;
 import com.github.kagkarlsson.jdbc.ResultSetMapper;
 import com.github.kagkarlsson.jdbc.SQLRuntimeException;
+import com.github.kagkarlsson.scheduler.ScheduledExecutionsFilter;
+import com.github.kagkarlsson.scheduler.SchedulerName;
+import com.github.kagkarlsson.scheduler.Serializer;
+import com.github.kagkarlsson.scheduler.TaskRepository;
+import com.github.kagkarlsson.scheduler.TaskResolver;
 import com.github.kagkarlsson.scheduler.TaskResolver.UnresolvedTask;
-import com.github.kagkarlsson.scheduler.jdbc.AndCondition;
-import com.github.kagkarlsson.scheduler.jdbc.AutodetectJdbcCustomization;
-import com.github.kagkarlsson.scheduler.jdbc.JdbcCustomization;
-import com.github.kagkarlsson.scheduler.jdbc.QueryBuilder;
 import com.github.kagkarlsson.scheduler.task.Execution;
 import com.github.kagkarlsson.scheduler.task.Task;
 import com.github.kagkarlsson.scheduler.task.TaskInstance;
@@ -450,7 +451,7 @@ public class JdbcTaskRepository implements TaskRepository {
         }
     }
 
-    public static class UnresolvedFilter implements AndCondition {
+    static class UnresolvedFilter implements AndCondition {
         private final List<UnresolvedTask> unresolved;
 
         public UnresolvedFilter(List<UnresolvedTask> unresolved) {
@@ -516,23 +517,4 @@ public class JdbcTaskRepository implements TaskRepository {
         }
     }
 
-    public static class JdbcTaskRepositoryContext {
-        public final TaskResolver taskResolver;
-        public final String tableName;
-        public final SchedulerName schedulerName;
-        public final JdbcRunner jdbcRunner;
-        public final Supplier<ResultSetMapper<List<Execution>>> resultSetMapper;
-
-        public JdbcTaskRepositoryContext(TaskResolver taskResolver,
-                                         String tableName,
-                                         SchedulerName schedulerName,
-                                         JdbcRunner jdbcRunner,
-                                         Supplier<ResultSetMapper<List<Execution>>> resultSetMapper) {
-            this.taskResolver = taskResolver;
-            this.tableName = tableName;
-            this.schedulerName = schedulerName;
-            this.jdbcRunner = jdbcRunner;
-            this.resultSetMapper = resultSetMapper;
-        }
-    }
 }
