+++ db-scheduler/src/test/java/com/github/kagkarlsson/scheduler/CustomTableNameTest.java
@@ -2,7 +2,7 @@ package com.github.kagkarlsson.scheduler;
 
 import com.github.kagkarlsson.jdbc.JdbcRunner;
 import com.github.kagkarlsson.jdbc.RowMapper;
-import com.github.kagkarlsson.scheduler.jdbc.DefaultJdbcCustomization;
+import com.github.kagkarlsson.scheduler.jdbc.JdbcTaskRepository;
 import com.github.kagkarlsson.scheduler.stats.StatsRegistry;
 import com.github.kagkarlsson.scheduler.task.Execution;
 import com.github.kagkarlsson.scheduler.task.Task;
